factor out mira and some stratification content
