+++ src/Stratify.jl
@@ -1,5 +1,5 @@
 module Stratify 
-export stratify, stratify_except, age_strata
+export stratify, stratify_except, age_strata, StrataSpec, add_cross_terms_with_rates
 
 using Catlab.CategoricalAlgebra
 using AlgebraicPetri
@@ -57,6 +57,43 @@ function stratify_except(pn1, pn2, type_system)
   end
   stratify(pn1[1]=>cross1,pn2[1]=>cross2, type_system)
 end
+
+
+#=
+THESE ARE STRATIFY FUNCTIONS FROM Structured-Epidemic-Modeling
+stratify(typed_model1, typed_model2) = ob(pullback(typed_model1, typed_model2))
+
+typed_stratify(typed_model1, typed_model2) =
+  compose(proj1(pullback(typed_model1, typed_model2)), typed_model1);
+=#
+
+abstract type AbstractStrataSpec end
+
+struct StrataSpec <: AbstractStrataSpec
+  tpn::ACSetTransformation
+  tlist::Vector{Vector{Symbol}}
+end
+
+function add_cross_terms_with_rates(ss::StrataSpec, type_system)
+  typed_pn = deepcopy(ss.tpn)
+  crossterms = deepcopy(ss.tlist)
+  pn = dom(typed_pn)
+  type_comps = Dict([k=>collect(v) for (k,v) in pairs(components(typed_pn))])
+  for (s_i,cts) in enumerate(crossterms)
+    for ct in cts 
+      type_ind = findfirst(==(ct), type_system[:tname])
+      is, os = [incident(type_system, type_ind, f) for f in [:it, :ot]]
+      new_t = isa(pn, LabelledReactionNet) ? add_part!(pn, :T; tname=ct, rate=1.0) : add_part!(pn, :T; tname=ct)
+      add_parts!(pn, :I, length(is); is=s_i, it=new_t)
+      add_parts!(pn, :O, length(os); os=s_i, ot=new_t)
+      push!(type_comps[:T], type_ind)
+      append!(type_comps[:I], is); append!(type_comps[:O], os); 
+    end
+  end
+  return homomorphism(pn, codom(typed_pn); initial=type_comps, 
+                        type_components=(Name=x->nothing,Rate=x->nothing, Concentration=x->nothing),)
+end
+
 
 """
 Create a petri net with `n` species, a unary transition on each, and a binary
